Issue 80: Move Doctrine fixtures in tests only
