+++ composer.json
@@ -44,6 +44,7 @@
     "autoload-dev": {
         "psr-4": {
             "App\\Tests\\Acceptance\\": "tests/acceptance/",
+            "App\\Tests\\Fixtures\\": "tests/fixtures/",
             "App\\Tests\\Integration\\": "tests/integration/",
             "App\\Tests\\System\\": "tests/system/"
         }
