+++ tests/acceptance/Context/FeatureContext.php
@@ -17,6 +17,7 @@ use Behat\Symfony2Extension\Context\KernelAwareContext;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\KernelInterface;
+use Webmozart\Assert\Assert;
 
 /**
  * @author Damien Carcel <damien.carcel@gmail.com>
@@ -50,14 +51,10 @@ class FeatureContext implements KernelAwareContext
     }
 
     /**
-     * @throws \RuntimeException
-     *
      * @Then a response should be received
      */
     public function aResponseShouldBeReceived(): void
     {
-        if (null === $this->response) {
-            throw new \RuntimeException('No response received');
-        }
+        Assert::notNull($this->response);
     }
 }
